Add files via upload
Since the instructions said we could use Chat GPT this is the code I got after a lot of tweaking. If either of your codes works better, I would prefer using that. I don't want to explain this madness in our presentation. Also do we want to make a cheating version.